+++ surface_morphometrics/__init__.py
@@ -4,4 +4,4 @@ See https://github.com/baradlab/surface_morphometrics and the `morphometrics`
 command line interface (defined in `surface_morphometrics.cli`).
 """
 
-__version__ = "2.0.0b1"
+__version__ = "2.0.0b3"